Use millisecond precision and UTC for timestamps (#13)
Because e.g. `2022-11-11T12:34:56.000Z` sorts _after_
`2022-11-11T12:34:56.000000001Z`, restrict the timestamps used with
SQLite to millisecond precision. This is because the built-in `strftime`
of SQLite only supports millisecond precision, otherwise we could have
switched to nanosecond precision everywhere.

An alternative is to not use the `strftime` function to create
timestamps, but then we lose the ability to set timestamps on e.g.
update triggers.

Also, always use UTC when (de)serializing `model.Time`.

Fixes #12.